+++ lib/src/model/content/blur_effect.dart
@@ -1,4 +1,4 @@
-import 'package:lottie/src/model/animatable/animatable_double_value.dart';
+import '../animatable/animatable_double_value.dart';
 
 class BlurEffect {
   final AnimatableDoubleValue blurriness;