+++ lib/src/model/layer/layer.dart
@@ -1,14 +1,13 @@
 import 'dart:ui';
-import 'package:lottie/src/model/content/blur_effect.dart';
-import 'package:lottie/src/model/content/drop_shadow_effect.dart';
-
 import '../../composition.dart';
 import '../../value/keyframe.dart';
 import '../animatable/animatable_double_value.dart';
 import '../animatable/animatable_text_frame.dart';
 import '../animatable/animatable_text_properties.dart';
 import '../animatable/animatable_transform.dart';
+import '../content/blur_effect.dart';
 import '../content/content_model.dart';
+import '../content/drop_shadow_effect.dart';
 import '../content/mask.dart';
 
 enum LayerType { preComp, solid, image, nullLayer, shape, text, unknown }
@@ -78,19 +77,32 @@ class Layer {
   }
 
   String toStringWithPrefix(String prefix) {
-    var sb = StringBuffer()..write(prefix)..write(name)..write('\n');
+    var sb = StringBuffer()
+      ..write(prefix)
+      ..write(name)
+      ..write('\n');
     var parent = composition.layerModelForId(parentId);
     if (parent != null) {
-      sb..write('\t\tParents: ')..write(parent.name);
+      sb
+        ..write('\t\tParents: ')
+        ..write(parent.name);
       parent = composition.layerModelForId(parent.parentId);
       while (parent != null) {
-        sb..write('->')..write(parent.name);
+        sb
+          ..write('->')
+          ..write(parent.name);
         parent = composition.layerModelForId(parent.parentId);
       }
-      sb..write(prefix)..write('\n');
+      sb
+        ..write(prefix)
+        ..write('\n');
     }
     if (masks.isNotEmpty) {
-      sb..write(prefix)..write('\tMasks: ')..write(masks.length)..write('\n');
+      sb
+        ..write(prefix)
+        ..write('\tMasks: ')
+        ..write(masks.length)
+        ..write('\n');
     }
     if (solidWidth != 0 && solidHeight != 0) {
       sb
@@ -99,9 +111,15 @@ class Layer {
         ..write('${solidWidth}x$solidHeight $solidColor');
     }
     if (shapes.isNotEmpty) {
-      sb..write(prefix)..write('\tShapes:\n');
+      sb
+        ..write(prefix)
+        ..write('\tShapes:\n');
       for (Object shape in shapes) {
-        sb..write(prefix)..write('\t\t')..write(shape)..write('\n');
+        sb
+          ..write(prefix)
+          ..write('\t\t')
+          ..write(shape)
+          ..write('\n');
       }
     }
     return sb.toString();