+++ lib/src/value/drop_shadow.dart
@@ -6,10 +6,36 @@ class DropShadow {
   final double distance;
   final double radius;
 
-  DropShadow({
+  const DropShadow({
     required this.color,
     required this.direction,
     required this.distance,
     required this.radius,
   });
+
+  DropShadow copyWith(
+      {Color? color, double? direction, double? distance, double? radius}) {
+    return DropShadow(
+      color: color ?? this.color,
+      direction: direction ?? this.direction,
+      distance: distance ?? this.distance,
+      radius: radius ?? this.radius,
+    );
+  }
+
+  @override
+  bool operator ==(other) {
+    return other is DropShadow &&
+        other.color == color &&
+        other.direction == direction &&
+        other.distance == distance &&
+        other.radius == radius;
+  }
+
+  @override
+  int get hashCode => hashValues(color, direction, distance, radius);
+
+  @override
+  String toString() => 'DropShadow(color: $color, direction: $direction, '
+      'distance: $distance, radius: $radius)';
 }